requirements.txt
Flask
gunicorn
firebase-admin
httpx
python-dotenv